todo: service side configuration changes, A/B setup
