+++ plugins/amazonq/codewhisperer/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codewhisperer/util/CodeWhispererFileCrawler.kt
@@ -5,35 +5,26 @@ package software.aws.toolkits.jetbrains.services.codewhisperer.util
 
 import com.intellij.openapi.application.runReadAction
 import com.intellij.openapi.fileEditor.FileEditorManager
+import com.intellij.openapi.fileEditor.impl.FileEditorManagerImpl
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.guessProjectDir
 import com.intellij.openapi.roots.TestSourcesFilter
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.psi.PsiDirectory
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiManager
 import software.aws.toolkits.core.utils.tryOrNull
+import software.aws.toolkits.jetbrains.isDeveloperMode
 import software.aws.toolkits.jetbrains.services.codewhisperer.model.ListUtgCandidateResult
+import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererConstants.CrossFile.NEIGHBOR_FILES_DISTANCE
 
 /**
  * An interface define how do we parse and fetch files provided a psi file or project
  * since different language has its own way importing other files or its own naming style for test file
  */
 interface FileCrawler {
-    /**
-     * parse the import statements provided a file
-     * @param psiFile of the file we are search with
-     * @return list of file reference from the import statements
-     */
-    suspend fun listFilesImported(psiFile: PsiFile): List<VirtualFile>
-
     fun listFilesUnderProjectRoot(project: Project): List<VirtualFile>
-
-    /**
-     * @param psiFile the file we are searching with, aka target file
-     * @return Files under the same package as the given file and exclude the given file
-     */
-    fun listFilesWithinSamePackage(psiFile: PsiFile): List<VirtualFile>
 
     /**
      * should be invoked at test files e.g. MainTest.java, or test_main.py
@@ -59,12 +50,9 @@ interface FileCrawler {
 }
 
 class NoOpFileCrawler : FileCrawler {
-    override suspend fun listFilesImported(psiFile: PsiFile): List<VirtualFile> = emptyList()
-
     override fun listFilesUnderProjectRoot(project: Project): List<VirtualFile> = emptyList()
-    override fun listUtgCandidate(target: PsiFile) = ListUtgCandidateResult(null, UtgStrategy.Empty)
 
-    override fun listFilesWithinSamePackage(psiFile: PsiFile): List<VirtualFile> = emptyList()
+    override fun listUtgCandidate(target: PsiFile) = ListUtgCandidateResult(null, UtgStrategy.Empty)
 
     override fun listCrossFileCandidate(target: PsiFile): List<VirtualFile> = emptyList()
 
@@ -99,18 +87,24 @@ abstract class CodeWhispererFileCrawler : FileCrawler {
         }
     }.orEmpty()
 
-    override fun listFilesWithinSamePackage(psiFile: PsiFile): List<VirtualFile> = runReadAction {
-        psiFile.containingDirectory?.files?.mapNotNull {
-            // exclude target file
-            if (it != psiFile) {
-                it.virtualFile
-            } else {
-                null
-            }
-        }.orEmpty()
+    override fun listCrossFileCandidate(target: PsiFile): List<VirtualFile> = if (isDeveloperMode()) {
+        val previousSelected = listPreviousSelectedFile(target)
+        val neighbors = neighborFiles(target).mapNotNull { it.virtualFile }
+        val result = previousSelected + neighbors
+        println("MRU list: ${previousSelected.map { it.name }}")
+        println("neighbors: ${neighbors.map { it.name }}")
+        result.distinctBy { it.name }
+    } else {
+        listAllOpenedFilesSortedByDist(target)
+    }.also {
+        val logStr = it.map { file -> file.name }
+        println("crossfile candidates: $logStr")
     }
 
-    override fun listCrossFileCandidate(target: PsiFile): List<VirtualFile> {
+    /**
+     * Default strategy will return all opened files sorted with file distance against the target
+     */
+    private fun listAllOpenedFilesSortedByDist(target: PsiFile): List<VirtualFile> {
         val targetFile = target.virtualFile
 
         val openedFiles = runReadAction {
@@ -128,6 +122,22 @@ abstract class CodeWhispererFileCrawler : FileCrawler {
         }
 
         return fileToFileDistanceList.sortedBy { it.second }.map { it.first }
+    }
+
+    /**
+     * New strategy will return opened files sorted by timeline the file is used (Most Recently Used), which should be as same as JB's file switcher (ctrl + tab)
+     * Note: test file is included here unlike the default strategy (thus different predicate inside filter)
+     */
+    private fun listPreviousSelectedFile(target: PsiFile): List<VirtualFile> {
+        val targetVFile = target.virtualFile
+        return runReadAction {
+            (FileEditorManager.getInstance(target.project) as FileEditorManagerImpl).getSelectionHistory()
+                .map { it.first }
+                .filter {
+                    it.name != targetVFile.name &&
+                        isSameDialect(it.extension)
+                }
+        }
     }
 
     override fun listUtgCandidate(target: PsiFile): ListUtgCandidateResult {
@@ -164,6 +174,71 @@ abstract class CodeWhispererFileCrawler : FileCrawler {
     private fun isSameDialect(fileExt: String?): Boolean = fileExt?.let {
         dialects.contains(fileExt)
     } ?: false
+
+    /**
+     *     1. A: root/util/context/a.ts
+     *     2. B: root/util/b.ts
+     *     3. C: root/util/service/c.ts
+     *     4. D: root/d.ts
+     *     5. E: root/util/context/e.ts
+     *     6. F: root/util/foo/bar/baz/f.ts
+     *
+     *   neighborfiles(A) = [B, E]
+     *   neighborfiles(B) = [A, C, D, E]
+     *   neighborfiles(C) = [B,]
+     *   neighborfiles(D) = [B,]
+     *   neighborfiles(E) = [A, B]
+     *   neighborfiles(F) = []
+     *
+     *      A B C D E F
+     *   A  x 1 2 2 0 4
+     *   B  1 x 1 1 1 3
+     *   C  2 1 x 2 2 4
+     *   D  2 1 2 x 2 4
+     *   E  0 1 2 2 x 4
+     *   F  4 3 4 4 4 x
+     */
+    fun neighborFiles(psiFile: PsiFile) = search(psiFile, NEIGHBOR_FILES_DISTANCE).filterNot { it == psiFile }.toSet()
+
+    private fun search(psiFile: PsiFile, distance: Int): Set<PsiFile> = runReadAction {
+        search(psiFile.containingDirectory, distance, true) +
+            search(psiFile.containingDirectory, distance, false)
+    }
+
+    private fun search(psiDir: PsiDirectory, distance: Int, goDown: Boolean): Set<PsiFile> {
+        var d = distance
+        val res = mutableListOf<PsiFile>()
+        val pendingVisit = mutableListOf(psiDir)
+        val visisted = mutableMapOf(psiDir to false)
+
+        while (d >= 0 && pendingVisit.isNotEmpty()) {
+            var toVisit = emptyList<PsiDirectory>()
+            for (dir in pendingVisit) {
+                if (visisted[dir] == true) {
+                    continue
+                }
+
+                val fs = dir.files
+                res.addAll(fs)
+
+                val dirs = if (goDown) {
+                    dir.subdirectories.toList()
+                } else {
+                    dir.parentDirectory?.let {
+                        listOf(it)
+                    } ?: emptyList()
+                }
+
+                toVisit = dirs
+                visisted[dir] = true
+            }
+
+            pendingVisit.addAll(toVisit)
+            d--
+        }
+
+        return res.toSet()
+    }
 
     companion object {
         // TODO: move to CodeWhispererUtils.kt
